# docs/swagger.yaml
definitions:
...
  models.User:
    properties:
      alias:
        type: string
      birthdate:
        example: YYYY-MM-DD
        type: string
      email:
        type: string
      followers:
        type: integer
      follows:
        type: integer
      fullname:
        type: string
      interests:
        items:
          type: string
        type: array
      is_admin:
        type: boolean
      nick:
        type: string
      ocupation:
        type: string
      pic:
        type: string
      uid:
        type: string
      zone:
        additionalProperties:
          type: number
        example:
          latitude: 0
          longitude: 0
        type: object
    type: object
  models.UserCreate:
    properties:
      alias:
        type: string
      birthdate:
        example: YYYY-MM-DD
        type: string
      email:
        type: string
      fullname:
        type: string
      interests:
        items:
          type: string
        type: array
      nick:
        type: string
      ocupation:
        type: string
      pic:
        type: string
      zone:
        additionalProperties:
          type: number
        example:
          latitude: 0
          longitude: 0
        type: object
    required:
    - alias
    - birthdate
    - email
    - fullname
    - interests
    - nick
    - ocupation
    - pic
    - zone
    type: object
  models.UserPublic:
    properties:
      alias:
        type: string
      followers:
        type: integer
      follows:
        type: integer
      interests:
        items:
          type: string
        type: array
      nick:
        type: string
      pic:
        type: string
      uid:
        type: string
    type: object
  models.UserUpdate:
    properties:
      alias:
        type: string
      interests:
        items:
          type: string
        type: array
      nick:
        type: string
      ocupation:
        type: string
      pic:
        type: string
      zone:
        additionalProperties:
          type: number
        example:
          latitude: 0
          longitude: 0
        type: object
    type: object
info:
  contact: {}
  title: SnapMsg API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - '*/*'
      parameters:
      - description: user id
        in: query
        name: uid
        type: string
      - description: user email
        in: query
        name: email
        type: string
      - description: user nickname
        in: query
        name: nick
        type: string
      - default: 100
        description: max results
        in: query
        maximum: 100
        minimum: 0
        name: limit
        required: true
        type: integer
      - default: 0
        description: page number
        in: query
        minimum: 0
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserPublic'
            type: array
      security:
      - Bearer: []
      summary: List public users data filtering by query
      tags:
      - users methods
      x-order: "1"
    post:
      consumes:
      - application/json
      parameters:
      - description: User creation data
        in: body
        name: userinfo
        required: true
        schema:
          $ref: '#/definitions/models.UserCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Create a user
      tags:
      - users methods
      x-order: "0"
  /users/{uid}/followers:
    get:
      consumes:
      - application/json
      parameters:
      - description: user id to list followers
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: List followers of given user
      tags:
      - users methods
  /users/{uid}/follows:
    get:
      consumes:
      - application/json
      parameters:
      - description: user id to list follows
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: List follows of a given user
      tags:
      - users methods
  /users/{uid}/follows/{otheruid}:
    get:
      consumes:
      - application/json
      parameters:
      - description: user id who follows
        in: path
        name: uid
        required: true
        type: string
      - description: the user id beeing followed
        in: path
        name: otheruid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - Bearer: []
      summary: Check if uid follows otheruid
      tags:
      - users methods
  /users/me:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete the user making the request
      tags:
      - users methods
    get:
      consumes:
      - '*/*'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - Bearer: []
      summary: Get private data of current user
      tags:
      - users methods
    patch:
      consumes:
      - application/json
      parameters:
      - description: data for update the user
        in: body
        name: UserUpdate
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Update current user data
      tags:
      - users methods
  /users/me/follows/{uid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: user id to unfollow
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Unfollow a given user
      tags:
      - users methods
    post:
      consumes:
      - application/json
      parameters:
      - description: user id to follow
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Follow a given user
      tags:
      - users methods
  /users/me/recommended:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserPublic'
            type: array
      security:
      - Bearer: []
      summary: Recommend users for the user making the request
      tags:
      - users methods
      x-order: "2"
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and Firebase JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"